Delete user account

+++ controllers/user/userController.js
@@ -2,6 +2,7 @@ const authenticateUser = require('../../middleware/authenticateUser')
 const User = require('../../models/user/userModel');
 const Expert = require('../../models/legalExpert/expertModel');
 const XUser = require('../../utils/constants/XUser');
+const bcrypt = require('bcrypt');
 
 /** Check if user exists **/
 findUser = async (req, res) => {
@@ -21,20 +22,29 @@ findUser = async (req, res) => {
 
 /**  Delete a user account - Cascade delete (profile, ask expert queries, user) **/
 deleteUserAccount = async (req, res) => {
-    try{
-    const userData = await authenticateUser(req, res);
-    const response = await User.deleteOne({ _id: userData.userId });
-    if (response.deletedCount == 0)
-        return res.status(404).json({ msg: 'User not found' });
-    else
-        res
-            .clearCookie('token', {
-                    httpOnly: true,
-                    secure: true,
-                    sameSite: 'None',
-                })
-            .status(200)
-            .json("Successfully deleted");
+    try {
+        const userData = await authenticateUser(req, res);
+        var userDoc = await User.findOne({ _id: userData.userId }, { password: 1 });
+        if (!userDoc) {
+            return res.status(404).json({ msg: 'User not found' });
+        }
+        var storedPassword = userDoc.password;
+        const comparePassword = await bcrypt.compare(req.body.password, storedPassword);
+        if (!comparePassword) {
+            return res.status(400).json({ msg: "Wrong password" });
+        }
+        else {
+            const response = await User.deleteOne({ _id: userData.userId });
+            if (response.deletedCount == 1)
+                res
+                    .clearCookie('token', {
+                        httpOnly: true,
+                        secure: true,
+                        sameSite: 'None',
+                    })
+                    .status(200)
+                    .json({msg: "Account deleted successfully"});
+        }
     }
     catch (error) {
         console.log(error);
